backend/refactor: decouple Price module
