+++ lib/mobility-core/mobility-core.cabal
@@ -283,6 +283,7 @@ library
       Kernel.Types.Logging
       Kernel.Types.MonadGuid
       Kernel.Types.Predicate
+      Kernel.Types.Price
       Kernel.Types.Registry
       Kernel.Types.Registry.API
       Kernel.Types.Registry.Routes
