+++ lib/mobility-core/src/Kernel/Beam/Lib/UtilsTH.hs
@@ -34,7 +34,6 @@ import qualified Database.PostgreSQL.Simple.FromField as DPSF
 import qualified EulerHS.KVConnector.Types as KV
 import EulerHS.Prelude hiding (Type, words)
 import Kernel.External.Encryption (DbHash (..))
-import Kernel.Types.Common ()
 import Kernel.Types.FromField (fromFieldEnum, fromFieldJSON)
 import Kernel.Utils.Text (encodeToText)
 import Language.Haskell.TH
